+++ pkg/main.go
@@ -14,7 +14,7 @@ func main() {
 	// to exit by itself using os.Exit. Manage automatically manages life cycle
 	// of datasource instances. It accepts datasource instance factory as first
 	// argument. This factory will be automatically called on incoming request
-	// from Grafana to create different instances of SampleDatasource (per datasource
+	// from Grafana to create different instances of Datasource (per datasource
 	// ID). When datasource configuration changed Dispose method will be called and
 	// new datasource instance created using NewSampleDatasource factory.
 	if err := datasource.Manage("databricks-community", plugin.NewSampleDatasource, datasource.ManageOpts{}); err != nil {