+++ src/datasource.ts
@@ -3,14 +3,16 @@ import {DataSourceWithBackend, getTemplateSrv} from '@grafana/runtime';
 import {MyDataSourceOptions, MyQuery} from './types';
 import {map, switchMap} from 'rxjs/operators';
 import {firstValueFrom} from 'rxjs';
-import {QuerrySuggestions} from "./QuerrySuggestions";
+import {QuerySuggestions} from "./components/Suggestions/QuerySuggestions";
 
 export class DataSource extends DataSourceWithBackend<MyQuery, MyDataSourceOptions> {
-    public suggestionProvider: QuerrySuggestions;
+    public suggestionProvider: QuerySuggestions;
+    public autoCompletionEnabled: boolean;
     constructor(instanceSettings: DataSourceInstanceSettings<MyDataSourceOptions>) {
         super(instanceSettings);
         this.annotations = {}
-        this.suggestionProvider = new QuerrySuggestions(this);
+        this.suggestionProvider = new QuerySuggestions(this);
+        this.autoCompletionEnabled = instanceSettings.jsonData.autoCompletion || false;
     }
 
     applyTemplateVariables(query: MyQuery, scopedVars: ScopedVars) {