+++ src/module.ts
@@ -1,8 +1,8 @@
 import { DataSourcePlugin } from '@grafana/data';
 import { DataSource } from './datasource';
-import { ConfigEditor } from './ConfigEditor';
-import { QueryEditor } from './QueryEditor';
-import { QueryEditorHelp } from './QueryEditorHelp';
+import { ConfigEditor } from './components/ConfigEditor/ConfigEditor';
+import { QueryEditor } from './components/QueryEditor/QueryEditor';
+import { QueryEditorHelp } from './components/QueryEditor/QueryEditorHelp';
 import { MyQuery, MyDataSourceOptions } from './types';
 
 export const plugin = new DataSourcePlugin<DataSource, MyQuery, MyDataSourceOptions>(DataSource)